move test_utils mod into utils folder
