+++ sway-lsp/src/utils/mod.rs
@@ -4,4 +4,6 @@ pub mod debug;
 pub mod markdown;
 pub(crate) mod markup;
 pub(crate) mod sync;
+#[cfg(test)]
+pub(crate) mod test;
 pub(crate) mod token;
